+++ src/redturtle/prenotazioni/tests/test_delete_booking_api.py
@@ -102,9 +102,6 @@ class TestDeleteBookingApi(unittest.TestCase):
             response = session.delete("{}/@booking".format(self.portal_url))
         else:
             response = session.delete("{}/@booking/{}".format(self.portal_url, uid))
-        # per le restapi il commit qui non ha senso, senza i test si rompono,
-        # ma è un caso, va spostato dove serve veramente, questo non è il suo posto
-        transaction.commit()
         return response
 
     def test_endpoint_raise_404_without_uid(self):
@@ -172,6 +169,8 @@ class TestDeleteBookingApi(unittest.TestCase):
 
         response = self.get_response(session=self.api_session_anon, uid=uid)
 
+        transaction.commit()
+
         self.assertEqual(response.status_code, 204)
         self.assertEqual(
             len(self.portal.portal_catalog.unrestrictedSearchResults(UID=uid)),
@@ -197,6 +196,8 @@ class TestDeleteBookingApi(unittest.TestCase):
 
         response = self.get_response(session=self.api_session_user, uid=uid)
 
+        transaction.commit()
+
         self.assertEqual(response.status_code, 204)
         self.assertEqual(
             len(self.portal.portal_catalog.unrestrictedSearchResults(UID=uid)),
@@ -214,7 +215,6 @@ class TestDeleteBookingApi(unittest.TestCase):
             }
         )
         uid = booking.UID()
-        transaction.commit()
 
         self.assertEqual(
             len(self.portal.portal_catalog.unrestrictedSearchResults(UID=uid)),
@@ -222,6 +222,7 @@ class TestDeleteBookingApi(unittest.TestCase):
         )
 
         response = self.get_response(session=self.api_session_user2, uid=uid)
+        transaction.commit()
 
         self.assertEqual(response.status_code, 401)
         self.assertEqual(