+++ src/redturtle/prenotazioni/upgrades.py
@@ -17,28 +17,28 @@ from plone.contentrules.engine.interfaces import IRuleStorage
 from zope.component import getUtility
 from zope.component import queryUtility
 
-from redturtle.prenotazioni.behaviors.booking_folder.notification_email.notification_email import (
+from redturtle.prenotazioni.behaviors.booking_folder.email import (
     notify_on_confirm_message_default_factory,
 )
-from redturtle.prenotazioni.behaviors.booking_folder.notification_email.notification_email import (
+from redturtle.prenotazioni.behaviors.booking_folder.email import (
     notify_on_confirm_subject_default_factory,
 )
-from redturtle.prenotazioni.behaviors.booking_folder.notification_email.notification_email import (
+from redturtle.prenotazioni.behaviors.booking_folder.email import (
     notify_on_move_message_default_factory,
 )
-from redturtle.prenotazioni.behaviors.booking_folder.notification_email.notification_email import (
+from redturtle.prenotazioni.behaviors.booking_folder.email import (
     notify_on_move_subject_default_factory,
 )
-from redturtle.prenotazioni.behaviors.booking_folder.notification_email.notification_email import (
+from redturtle.prenotazioni.behaviors.booking_folder.email import (
     notify_on_refuse_message_default_factory,
 )
-from redturtle.prenotazioni.behaviors.booking_folder.notification_email.notification_email import (
+from redturtle.prenotazioni.behaviors.booking_folder.email import (
     notify_on_refuse_subject_default_factory,
 )
-from redturtle.prenotazioni.behaviors.booking_folder.notification_email.notification_email import (
+from redturtle.prenotazioni.behaviors.booking_folder.email import (
     notify_on_submit_message_default_factory,
 )
-from redturtle.prenotazioni.behaviors.booking_folder.notification_email.notification_email import (
+from redturtle.prenotazioni.behaviors.booking_folder.email import (
     notify_on_submit_subject_default_factory,
 )
 from redturtle.prenotazioni.events.prenotazione import set_booking_code