+++ CHANGES.rst
@@ -8,6 +8,9 @@ Changelog
 - Split booking notification gateways to 'App IO', 'Email' and 'SMS'
   [folix-01]
 
+- remove unused contentrules
+  [mamico]
+
 - Add booking reminder emails.
   [folix-01]
 
@@ -355,8 +358,9 @@ Changelog
 
 - Handle contentrules by the plone events and do not use contentrules anymore.
   [folix-01]
+
 - Change "day" type in week_table (TODO: need an upgrade step?).
-  [mauro]
+  [mamico]
 
 2.0.0.dev2 (2023-06-30)
 -----------------------